refactor(c3-01): extract duplicated logic into @llmwiki/shared
Move business logic from CLI commands and VS Code extension into shared library:

- Create packages/shared/src/utils.ts: slugify(), excerpt()
- Create packages/shared/src/search.ts: countOccurrences()
- Create packages/shared/src/ingest.ts: IngestResult, ingestSource()
- Create packages/shared/src/query.ts: QueryResult, QueryOutput, slugifyQuery(), queryWiki()
- Create packages/shared/src/status.ts: StatusResult, getWikiStatus()
- Update shared barrel to export all new modules
- CLI commands become thin wrappers with re-exports for backward compat
- VS Code commands.ts reduced by ~165 lines (uses shared functions)
- VS Code statusBar.ts uses getWikiStatus from shared
- Update VS Code test mocks for new shared function imports
- All 307 tests pass, lint clean, build successful
