+++ packages/shared/src/index.ts
@@ -4,3 +4,13 @@ export * from './log.js';
 export * from './sources.js';
 export * from './backlinks.js';
 export * from './lint.js';
+export * from './utils.js';
+export * from './search.js';
+export * from './ingest.js';
+export * from './query.js';
+export * from './status.js';
+export * from './utils.js';
+export * from './search.js';
+export * from './ingest.js';
+export * from './query.js';
+export * from './status.js';
