+++ packages/shared/src/search.ts
@@ -1,0 +1,15 @@
+/**
+ * Count occurrences of a term in text (case-insensitive).
+ */
+export function countOccurrences(text: string, term: string): number {
+  const lower = text.toLowerCase();
+  const t = term.toLowerCase();
+  if (!t) return 0;
+  let count = 0;
+  let pos = 0;
+  while ((pos = lower.indexOf(t, pos)) !== -1) {
+    count++;
+    pos += t.length;
+  }
+  return count;
+}
