+++ packages/vscode/src/commands.ts
@@ -1,16 +1,14 @@
 import * as vscode from 'vscode';
-import { join, resolve, basename, extname, relative } from 'node:path';
-import { readFile, writeFile, readdir, mkdir, stat } from 'node:fs/promises';
+import { join, resolve } from 'node:path';
+import { writeFile, mkdir } from 'node:fs/promises';
 import {
   readIndex,
-  readPage,
-  writePage,
-  listPages,
-  readLog,
-  addEntry,
-  appendEntry,
   directoryExists,
   lintWiki,
+  appendEntry,
+  ingestSource,
+  queryWiki,
+  getWikiStatus,
   type IndexEntry,
 } from '@llmwiki/shared';
 import type { WikiPagesTreeDataProvider } from './wikiPagesTree';
@@ -111,58 +109,14 @@ export function registerCommands(
     if (!selected || selected.length === 0) return;
 
     const sourcePath = selected[0].fsPath;
-    const root = resolve(workspaceFolder);
+    const result = await ingestSource(sourcePath, workspaceFolder, false);
 
-    // Path traversal check
-    const normalizedSource = resolve(sourcePath).replace(/\\/g, '/');
-    const normalizedRoot = root.replace(/\\/g, '/');
-    if (!normalizedSource.startsWith(normalizedRoot + '/')) {
-      vscode.window.showErrorMessage('Source file must be inside the project folder.');
+    if (result.status === 'error') {
+      vscode.window.showErrorMessage(result.error ?? 'Ingest failed.');
       return;
     }
 
-    const sourceContent = await readFile(sourcePath, 'utf-8');
-    const sourceStat = await stat(sourcePath);
-    const sourceFilename = basename(sourcePath);
-    const sourceExt = extname(sourcePath);
-    const slug = sourceFilename.replace(/\.[^.]+$/, '').toLowerCase()
-      .replace(/[^a-z0-9]+/g, '-').replace(/^-+|-+$/g, '');
-
-    const summaryRelPath = `sources/${slug}-summary.md`;
-    const summaryFullPath = join(wikiDir, summaryRelPath);
-    const relativeSourcePath = relative(root, resolve(sourcePath)).replace(/\\/g, '/');
-    const today = new Date().toISOString().slice(0, 10);
-
-    const excerpt = sourceContent.length > 500
-      ? sourceContent.slice(0, 500) + '…'
-      : sourceContent;
-
-    await writePage(summaryFullPath, {
-      frontmatter: {
-        type: 'source',
-        title: sourceFilename,
-        source_path: relativeSourcePath,
-        ingested: today,
-        tags: [],
-      },
-      body: `# ${sourceFilename}\n\n**Source:** ${relativeSourcePath}  \n**Type:** ${sourceExt || 'unknown'}  \n**Size:** ${sourceStat.size} bytes  \n**Ingested:** ${today}\n\n## Content Preview\n\n${excerpt}`,
-    });
-
-    await addEntry(indexPath, {
-      path: summaryRelPath,
-      title: sourceFilename,
-      summary: `Source file (${sourceExt || 'unknown'})`,
-      category: 'Sources',
-      tags: [],
-    });
-
-    await appendEntry(logPath, {
-      verb: 'ingested',
-      subject: sourceFilename,
-      details: `Ingested source "${sourceFilename}" → ${summaryRelPath}`,
-    });
-
-    vscode.window.showInformationMessage(`Ingested: ${summaryRelPath}`);
+    vscode.window.showInformationMessage(`Ingested: ${result.pages_created.join(', ')}`);
     providers.wikiPages.refresh();
   });
 
@@ -179,62 +133,14 @@ export function registerCommands(
     });
     if (!queryStr) return;
 
-    const entries = await readIndex(indexPath);
-    const terms = queryStr.toLowerCase().split(/\s+/).filter(Boolean);
-    if (terms.length === 0) return;
+    const output = await queryWiki(queryStr, workspaceFolder, false);
 
-    const countOccurrences = (text: string, term: string): number => {
-      const lower = text.toLowerCase();
-      if (!term) return 0;
-      let count = 0;
-      let pos = 0;
-      while ((pos = lower.indexOf(term, pos)) !== -1) {
-        count++;
-        pos += term.length;
-      }
-      return count;
-    };
-
-    interface ScoredResult {
-      title: string;
-      path: string;
-      score: number;
-      excerpt: string;
-    }
-
-    const results: ScoredResult[] = [];
-    for (const entry of entries) {
-      let score = 0;
-      for (const term of terms) {
-        score += countOccurrences(entry.title, term) * 3;
-        score += countOccurrences(entry.summary, term) * 2;
-      }
-      if (score === 0) continue;
-
-      let body = '';
-      try {
-        const page = await readPage(join(wikiDir, entry.path));
-        body = page.body;
-        for (const term of terms) {
-          score += countOccurrences(body, term);
-        }
-      } catch {
-        // Page file missing — use index score only
-      }
-
-      const cleaned = body.replace(/\s+/g, ' ').trim();
-      const excerpt = cleaned.length > 200 ? cleaned.slice(0, 200) + '…' : cleaned;
-      results.push({ title: entry.title, path: entry.path, score, excerpt });
-    }
-
-    results.sort((a, b) => b.score - a.score);
-
-    if (results.length === 0) {
+    if (output.matches === 0) {
       vscode.window.showInformationMessage(`No results for "${queryStr}".`);
       return;
     }
 
-    const items = results.map((r) => ({
+    const items = output.results.map((r) => ({
       label: r.title,
       description: `Score: ${r.score}`,
       detail: r.excerpt,
@@ -242,7 +148,7 @@ export function registerCommands(
     }));
 
     const pick = await vscode.window.showQuickPick(items, {
-      placeHolder: `${results.length} result(s) for "${queryStr}"`,
+      placeHolder: `${output.matches} result(s) for "${queryStr}"`,
       matchOnDetail: true,
     });
     if (pick) {
@@ -279,45 +185,14 @@ export function registerCommands(
 
   // ── llmwiki.status ───────────────────────────────────────────
   reg('llmwiki.status', async () => {
-    let sourceCount = 0;
-    try {
-      const rawEntries = await readdir(rawDir, { withFileTypes: true, recursive: true });
-      sourceCount = rawEntries.filter((e) => e.isFile()).length;
-    } catch {
-      // raw/ doesn't exist
-    }
-
-    const allPages = await listPages(wikiDir);
-    const wikiPages = allPages.filter((p) => {
-      const rel = relative(wikiDir, p).replace(/\\/g, '/');
-      return rel !== 'index.md' && rel !== 'log.md';
-    });
-    const wikiPageCount = wikiPages.length;
-
-    const logEntries = await readLog(logPath);
-    const ingestEntries = logEntries.filter((e) => e.verb.toLowerCase().includes('ingest'));
-    const lastIngestDate = ingestEntries.length > 0
-      ? ingestEntries[ingestEntries.length - 1].date
-      : null;
-
-    const indexEntries = await readIndex(indexPath);
-    const indexedPaths = new Set(indexEntries.map((e) => e.path));
-    const orphanPageCount = wikiPages.filter((p) => {
-      const rel = relative(wikiDir, p).replace(/\\/g, '/');
-      return !indexedPaths.has(rel);
-    }).length;
-
-    const indexedPageCount = wikiPageCount - orphanPageCount;
-    const coveragePct = wikiPageCount > 0
-      ? Math.round((indexedPageCount / wikiPageCount) * 100)
-      : 100;
+    const status = await getWikiStatus(workspaceFolder);
 
     const lines = [
-      `Pages: ${wikiPageCount}`,
-      `Sources: ${sourceCount}`,
-      `Coverage: ${coveragePct}%`,
-      `Last ingest: ${lastIngestDate ?? '—'}`,
-      `Orphans: ${orphanPageCount}`,
+      `Pages: ${status.wiki_page_count}`,
+      `Sources: ${status.source_count}`,
+      `Coverage: ${status.index_coverage_pct}%`,
+      `Last ingest: ${status.last_ingest_date ?? '—'}`,
+      `Orphans: ${status.orphan_page_count}`,
     ];
     vscode.window.showInformationMessage(`Wiki Status — ${lines.join(' | ')}`);
   });
